+++ core/src/api/types.rs
@@ -8,19 +8,20 @@ use avail_rust::{
 	AvailHeader, H256,
 };
 use codec::Encode;
-use color_eyre::{
-	eyre::{eyre, WrapErr},
-	Report, Result,
-};
+#[cfg(not(target_arch = "wasm32"))]
+use color_eyre::eyre::eyre;
+use color_eyre::{eyre::WrapErr, Report, Result};
 use derive_more::From;
+#[cfg(not(target_arch = "wasm32"))]
 use hyper::{http, StatusCode};
 use serde::{de, Deserialize, Deserializer, Serialize, Serializer};
-use std::{
-	collections::{HashMap, HashSet},
-	sync::Arc,
-};
+use std::collections::HashSet;
+#[cfg(not(target_arch = "wasm32"))]
+use std::{collections::HashMap, sync::Arc};
+#[cfg(not(target_arch = "wasm32"))]
 use tokio::sync::{mpsc::UnboundedSender, RwLock};
 use uuid::Uuid;
+#[cfg(not(target_arch = "wasm32"))]
 use warp::{
 	ws::{self, Message},
 	Reply,
@@ -41,6 +42,7 @@ use crate::{
 #[derive(Debug)]
 pub struct InternalServerError {}
 
+#[cfg(not(target_arch = "wasm32"))]
 impl warp::reject::Reject for InternalServerError {}
 
 #[derive(Serialize, Deserialize, Clone, Debug)]
@@ -49,6 +51,7 @@ pub struct Version {
 	pub network_version: String,
 }
 
+#[cfg(not(target_arch = "wasm32"))]
 impl Reply for Version {
 	fn into_response(self) -> warp::reply::Response {
 		warp::reply::json(&self).into_response()
@@ -124,6 +127,7 @@ pub struct SubmitResponse {
 	pub index: u32,
 }
 
+#[cfg(not(target_arch = "wasm32"))]
 impl Reply for SubmitResponse {
 	fn into_response(self) -> warp::reply::Response {
 		warp::reply::json(&self).into_response()
@@ -174,6 +178,7 @@ impl From<&SharedConfig> for Vec<Mode> {
 	}
 }
 
+#[cfg(not(target_arch = "wasm32"))]
 impl Reply for Status {
 	fn into_response(self) -> warp::reply::Response {
 		warp::reply::json(&self).into_response()
@@ -318,6 +323,7 @@ impl Block {
 	}
 }
 
+#[cfg(not(target_arch = "wasm32"))]
 impl Reply for Block {
 	fn into_response(self) -> warp::reply::Response {
 		warp::reply::json(&self).into_response()
@@ -348,6 +354,7 @@ pub struct Header {
 	received_at: u64,
 }
 
+#[cfg(not(target_arch = "wasm32"))]
 impl Reply for Header {
 	fn into_response(self) -> warp::reply::Response {
 		warp::reply::json(&self).into_response()
@@ -547,6 +554,7 @@ pub struct DataResponse {
 	pub data_transactions: Vec<DataTransaction>,
 }
 
+#[cfg(not(target_arch = "wasm32"))]
 impl Reply for DataResponse {
 	fn into_response(self) -> warp::reply::Response {
 		warp::reply::json(&self).into_response()
@@ -618,6 +626,7 @@ pub enum PublishMessage {
 	DataVerified(DataMessage),
 }
 
+#[cfg(not(target_arch = "wasm32"))]
 impl PublishMessage {
 	fn apply_filter(&mut self, fields: &HashSet<DataField>) {
 		match self {
@@ -630,6 +639,7 @@ impl PublishMessage {
 	}
 }
 
+#[cfg(not(target_arch = "wasm32"))]
 impl TryFrom<PublishMessage> for Message {
 	type Error = Report;
 	fn try_from(value: PublishMessage) -> Result<Self, Self::Error> {
@@ -639,13 +649,16 @@ impl TryFrom<PublishMessage> for Message {
 	}
 }
 
+#[cfg(not(target_arch = "wasm32"))]
 pub type Sender = UnboundedSender<Result<ws::Message, warp::Error>>;
 
+#[cfg(not(target_arch = "wasm32"))]
 pub struct WsClient {
 	pub subscription: Subscription,
 	pub sender: Option<Sender>,
 }
 
+#[cfg(not(target_arch = "wasm32"))]
 impl WsClient {
 	pub fn new(subscription: Subscription) -> Self {
 		WsClient {
@@ -665,9 +678,11 @@ impl WsClient {
 	}
 }
 
+#[cfg(not(target_arch = "wasm32"))]
 #[derive(Clone)]
 pub struct WsClients(pub Arc<RwLock<HashMap<String, WsClient>>>);
 
+#[cfg(not(target_arch = "wasm32"))]
 impl WsClients {
 	pub async fn set_sender(&self, subscription_id: &str, sender: Sender) -> Result<()> {
 		let mut clients = self.0.write().await;
@@ -707,6 +722,7 @@ impl WsClients {
 	}
 }
 
+#[cfg(not(target_arch = "wasm32"))]
 impl Default for WsClients {
 	fn default() -> Self {
 		Self(Arc::new(RwLock::new(HashMap::new())))
@@ -718,6 +734,7 @@ pub struct SubscriptionId {
 	pub subscription_id: String,
 }
 
+#[cfg(not(target_arch = "wasm32"))]
 impl Reply for SubscriptionId {
 	fn into_response(self) -> warp::reply::Response {
 		warp::reply::json(&self).into_response()
@@ -754,6 +771,7 @@ impl<T> Response<T> {
 	}
 }
 
+#[cfg(not(target_arch = "wasm32"))]
 impl TryFrom<ws::Message> for Request {
 	type Error = Report;
 
@@ -816,6 +834,7 @@ impl Error {
 		Self::new(Some(request_id), None, ErrorCode::BadRequest, message)
 	}
 
+	#[cfg(not(target_arch = "wasm32"))]
 	fn status(&self) -> StatusCode {
 		match self.error_code {
 			ErrorCode::NotFound => StatusCode::NOT_FOUND,
@@ -825,6 +844,7 @@ impl Error {
 	}
 }
 
+#[cfg(not(target_arch = "wasm32"))]
 impl Reply for Error {
 	fn into_response(self) -> warp::reply::Response {
 		http::Response::builder()