+++ core/CHANGELOG.md
@@ -1,6 +1,20 @@
 # Changelog
 
-## 1.2.0
+## 1.2.3
+
+- Update `rust-libp2p` and related libraries to `0.55`
+
+## [1.2.2](https://github.com/availproject/avail-light/tree/avail-light-core-v1.2.2) - 2025-02-20
+
+- Support v2 API in WASM
+- Update avail-rust dependency to v0.1.0
+
+## [1.2.1](https://github.com/availproject/avail-light/tree/avail-light-core-v1.2.1) - 2025-02-10
+
+- Batch rows RPC requests
+- Remove UPNP behaviour
+
+## [1.2.0](https://github.com/availproject/avail-light/tree/avail-light-core-v1.2.0) - 2025-02-06
 
 - Remove multiaddress metric attribute
 - Add timestamp to v2 header API